AndyMe.Funds
AndmeFunds is a crowdfunding or call it a brige that joints people in the world through care love and support to who needs it most. Inside the heart of Andrew is full of helping but he can't help the whole world by himself but with togetherness we can touch and change the world. It's with ONE!